Update README.md
WhosTagged? is a legal mod that adds a UI panel to the game that shows players who are tagged, and players that are not. It also shows how close the lava monkey is from you hy displaying: Really Close, Close, Or Far, by the players name. BTW This is a mod, not a cheat, meaning it will only work in modded lobbies, requirements to work are either Utilla or Newtilla.